feat(theme): improve Redoc component to display it in full size when used in MDX file
